auth

+++ routes/lostitemRoutes.js
@@ -1,9 +1,11 @@
 const express = require('express');
 const router = express.Router();
 const LostItem = require('../models/lostitem');
+const User = require('../models/User');
+const auth = require('./../middlewares/auth').auth
 
 // GET all lost items
-router.get('/', async (req, res) => {
+router.get('/', auth,  async (req, res) => {
   try {
     const items = await LostItem.find();
     res.json(items);
@@ -13,7 +15,7 @@ router.get('/', async (req, res) => {
 });
 
 // POST a new lost item
-router.post('/', async (req, res) => {
+router.post('/', auth, async (req, res) => {
   const lostItem = new LostItem(req.body);
   try {
     const newItem = await lostItem.save();
@@ -24,7 +26,7 @@ router.post('/', async (req, res) => {
 });
 
 // GET lost item by ID (MongoDB _id)
-router.get('/:id', async (req, res) => {
+router.get('/:id', auth,  async (req, res) => {
   try {
     const item = await LostItem.findById(req.params.id);
     if (!item) {
@@ -37,8 +39,11 @@ router.get('/:id', async (req, res) => {
 });
 
 // DELETE lost item by ID
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', auth, async (req, res) => {
   try {
+    let user = await User.findOne({email: req.user.email})
+    if(user ===null){ return res.status(400).json({message:"Invalid User!"})}
+    if(user.role === 'user'){ return res.status(403).json({message:"Access Denied!"})}
     const deletedItem = await LostItem.findByIdAndDelete(req.params.id);
     if (!deletedItem) {
       return res.status(404).json({ message: 'Lost item not found' });
@@ -50,7 +55,7 @@ router.delete('/:id', async (req, res) => {
 });
 
 // PUT (update) lost item by ID
-router.put('/:id', async (req, res) => {
+router.put('/:id', auth,  async (req, res) => {
   try {
     const item = await LostItem.findById(req.params.id);
     if (!item) return res.status(404).json({ error: 'Lost item not found' });
